#1114 WIP: Added field in Sim Well Collection, renamed a bit as preparations

+++ ApplicationCode/ProjectDataModel/RimWellLogPlot.cpp
@@ -422,7 +422,7 @@ QString RimWellLogPlot::asciiDataForPlotExport() const
         }
 
         
-        for (int i = curveDepths.size() -1; i >= 0; i--)
+        for (int i = static_cast<int>( curveDepths.size()) - 1; i >= 0; i--)
         {
             if (i == curveDepths.size() - 1)
             {
